refactor(frontend): вынос favicon-подсистемы из stores/theme.ts
Этап F3: динамический favicon (перекрашивание /favicon.svg в акцент
темы, localStorage-кэш, DOM-линки) выделен из theme.ts (845 строк) в
themeFavicon.ts (181 строка); тема вызывает refreshDynamicFavicon/
applyCachedDynamicFavicon при применении состояния.

Цикл модулей theme ↔ themeFavicon безопасен: favicon импортирует из
theme только type ThemeTokenMap и hoisted function-декларации
(normalizeHexColor) + DEFAULT_CUSTOM_THEME, ни одна не вызывается на
этапе инициализации модуля.

Эквивалентность верифицирована мультимножествами: расхождения — только
export-модификаторы и import-блоки. svelte-check 0/0, eslint 0,
vitest 1343, build зелёный.
